Hide HashingOutputStream
This is not used from the outside anymore
